% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{foreign}
\alias{foreign}
\alias{exportQuodata}
\alias{exportSAS}
\alias{exportExcelMacro}
\title{Support Other Platforms}
\usage{
exportQuodata(obj)

exportSAS(obj)

exportExcelMacro(dest)
}
\arguments{
\item{obj}{A list returned by \code{\link{analyzeSingleLab}}.}

\item{dest}{The path to write the excel makro to.}
}
\value{
Nothing is returned.
}
\description{
Export formatted data or code for use by other platforms
}
\details{
The output of exportQuodata can be used on the QuoData website (<http://quodata.de/content/validation-qualitative-pcr-methods-single-laboratory>).
}
\examples{
x <- cbind(
 X=c( 0.1,1,2,5,10,20 ),
 S=c( 0,5,6,6,6,6 ),
 N=c( 6,6,6,6,6,6 )
)
obj <- analyzeSingleLab(x=x)
exportQuodata(obj)


}
