% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPOD.R
\name{plotPOD}
\alias{plotPOD}
\title{Generate Plot to Analyze Single Lab PCR Outcomes}
\usage{
plotPOD(obj, model = "auto", qLOD = 95, show.warnings = FALSE)
}
\arguments{
\item{obj}{A list returned by \code{\link{analyzeSingleLab}}.}

\item{model}{Simple or full model}

\item{qLOD}{The quantile(s) for LOD to be shown in the plot. Multiplied by \eqn{100} if less than one.}

\item{show.warnings}{Show the warning regarding significant deviation from \eqn{1} in the plot.}
}
\value{
The passed list 'obj' is returned invisibly.
}
\description{
Show POD curve and LOD value to validate qualitative PCR methods of a single laboratory.
}
\details{
If model is set to "auto", a plausiblity test is applied to determine if the POD curve bases on the simplified or on full parameter estimation. If the corrective parameter determined from the full model significantly differs from \eqn{1}, a message is shown in the plot. Testing for significant deviation is currently done by checking the condition \eqn{1-b>0.2}. The threshould \eqn{0.2} has been determined empirically to agree with the original webtool and might be changed in future versions of the package.

Three cases can be distinguished. First, the value for the slope parameter b is significantly less than \eqn{1}. This means the average amplification probability is higher at higher dilution levels than at lower dilution levels. Such a situation can be related to: inhibitory matrix effects, a large variability in the amplification process from the one test to another under repeatability conditions, or accidental problems causing false positives if the number of copies of the target DNA sequence is less than \eqn{1}. Second, the calculated POD curve indicates sensitivity better than achievable according to the theoretical POD curve. Third, the number of positive test results is significantly higher than expected at nominal copies of nominal DNA concentrations in \eqn{[0.5,1.5]}. In this case check the correctness of the serial dilution.

Another warning appears if the LOD of interest exceeds the highest number of considered nominal copies.
}
\examples{
x <- cbind(
 X=c(0.1,1,2,5,10,20),
 S=c( 0,5,6,6,6,6 ),
 N=c( 6,6,6,6,6,6 )
)
obj <- analyzeSingleLab(x=x)
plotPOD(obj)
}
