% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPod.R
\name{print.pod}
\alias{print.pod}
\title{Summary of POD objects}
\usage{
\method{print}{pod}(x, ...)
}
\arguments{
\item{x}{An object of class 'pod'}

\item{...}{Other parameters, not supported yet.}
}
\value{
Nothing is returned.
}
\description{
Generate nicely formatted output of the POD object
}
\examples{
x <- cbind(
 X=c( 0.1,1,2,5,10,20 ),
 S=c( 0,5,6,6,6,6 ),
 N=c( 6,6,6,6,6,6 )
)
obj <- analyzeSingleLab(x=x)
print(obj)
# or just
obj

obj <- analyzeSingleLab(x=x, qLOD=c(50, 70, 95))
obj
}
